Persist database volumes

+++ server/run.sh
@@ -1,5 +1,12 @@
 DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
-docker run -tid --rm -p 5432:5432 --name hoc-db dhoodlum/havenofcode-db
+
+# keep the data volume running for persistence
+DB_CONTAINER_NAME="hoc-db"
+CID=$(docker ps -q -f status=running -f name=$DB_CONTAINER_NAME)
+if [ ! "${CID}" ]; then
+  docker run -tid --rm -p 5432:5432 --name hoc-db dhoodlum/havenofcode-db
+fi
+unset CID
 
 docker run -ti \
   --link hoc-db \
@@ -12,6 +19,3 @@ docker run -ti \
   -p 80:80 \
   dhoodlum/havenofcode-http \
   zsh
-
-echo 'stopping hoc-db'
-docker stop hoc-db
